Implementando autenticação, autorização e ajustando o codigo

+++ src/main/java/com/felipeduarte/services/UserService.java
@@ -6,11 +6,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort.Direction;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import com.felipeduarte.exceptions.AuthorizationException;
 import com.felipeduarte.models.User;
 import com.felipeduarte.models.dtos.UserDTO;
+import com.felipeduarte.models.dtos.UserDTOWithPassword;
 import com.felipeduarte.repositories.UserRepository;
+import com.felipeduarte.security.UserDetailsAdapter;
+import com.felipeduarte.security.UserDetailsServiceAdapter;
 
 @Service
 public class UserService {
@@ -18,11 +23,18 @@ public class UserService {
 	@Autowired
 	private UserRepository userRepository;
 	
-	public User save(User user) {
+	@Autowired
+	private BCryptPasswordEncoder bCryptPasswordEnconder;
+	
+	public User save(UserDTOWithPassword userDTOWithPassword) {
 		
-		Optional<User> optionalUser = this.userRepository.findByEmail(user.getEmail());
+		Optional<User> optionalUser = this.userRepository.findByEmail(userDTOWithPassword.getEmail());
 		
 		if(optionalUser.isPresent()) return null;
+		
+		User user = User.convertUserDTOWithPasswordToUser(userDTOWithPassword);
+		
+		user.setPassword(this.bCryptPasswordEnconder.encode(user.getPassword()));
 		
 		user = this.userRepository.save(user);
 		
@@ -43,6 +55,8 @@ public class UserService {
 			return userDTO;
 		}
 		
+		UserServicePermission.verifyPermissionUserUpdate(userDTO.getId());
+		
 		User user = User.convertUserDTOToUser(userDTO);
 		
 		user.setPassword(optionalUser.get().getPassword());
@@ -56,20 +70,24 @@ public class UserService {
 	
 	public boolean delete(Long id) {
 		
-		User user = this.findById(id);
+		User user = this.findById(id,true);
 		
 		if(user == null) return false;
+		
+		UserServicePermission.verifyPermissionUserDelete(id);
 		
 		this.userRepository.delete(user);
 		
 		return true;
 	}
 	
-	public User findById(Long id) {
+	public User findById(Long id, boolean verificaPermissao) {
 		
 		Optional<User> optionalUser = this.userRepository.findById(id);
 		
 		if(!optionalUser.isPresent()) return null;
+		
+		if(verificaPermissao) UserServicePermission.verifyUserId(id);
 		
 		return optionalUser.get();
 	}
@@ -96,6 +114,40 @@ public class UserService {
 		Page<User> pageUser = this.userRepository.findAll(page);
 		
 		return pageUser;
+	}
+	
+	private static class UserServicePermission{
+		
+		public static void verifyPermissionUserUpdate(Long id){
+			
+			UserDetailsAdapter user = UserDetailsServiceAdapter.getUser();
+			
+			if(user == null || !user.getId().equals(id))
+				throw new AuthorizationException("Usuario não tem permissão para "
+						+ "alterar dados dos outros usuarios");
+		}
+		
+		public static void verifyPermissionUserDelete(Long id) {
+			
+			UserDetailsAdapter user = UserDetailsServiceAdapter.getUser();
+			
+			if(user == null || (user.hasRole("ROLE_USER") && !user.getId().equals(id)) ) {
+				throw new AuthorizationException("Usuário não tem permissão para" 
+					+ "excluir outros usuários");
+			}
+		
+		}
+		
+		public static void verifyUserId(Long id) {
+			
+			UserDetailsAdapter user = UserDetailsServiceAdapter.getUser();
+			
+			if(user == null) throw new AuthorizationException("Usuario não logado!");
+			
+			if(!user.getId().equals(id)) throw new AuthorizationException("Este id não pertence ao usuario logado!");
+			
+		}
+		
 	}
 	
 }
